requirements.txt
fastapi==0.63.0
uvicorn==0.13.3
pydantic[email]
motor==2.2.0
python-decouple==3.4
